refactor: Remove the interfaces for publishing events.
