--- src/main/java/com/synopsys/integration/detect/lifecycle/shutdown/DefaultExitCodePublisher.java
@@ -1,24 +1,0 @@
-/*
- * synopsys-detect
- *
- * Copyright (c) 2021 Synopsys, Inc.
- *
- * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
- */
-package com.synopsys.integration.detect.lifecycle.shutdown;
-
-import com.synopsys.integration.detect.workflow.event.Event;
-import com.synopsys.integration.detect.workflow.event.EventSystem;
-
-public class DefaultExitCodePublisher implements ExitCodePublisher {
-    private EventSystem eventSystem;
-
-    public DefaultExitCodePublisher(EventSystem eventSystem) {
-        this.eventSystem = eventSystem;
-    }
-
-    @Override
-    public void publishExitCode(ExitCodeRequest exitCodeRequest) {
-        eventSystem.publishEvent(Event.ExitCode, exitCodeRequest);
-    }
-}
