+++ src/main/java/com/synopsys/integration/detect/workflow/status/StatusEventPublisher.java
@@ -7,12 +7,26 @@
  */
 package com.synopsys.integration.detect.workflow.status;
 
+import com.synopsys.integration.detect.workflow.event.Event;
+import com.synopsys.integration.detect.workflow.event.EventSystem;
 import com.synopsys.integration.detect.workflow.result.DetectResult;
 
-public interface StatusEventPublisher {
-    void publishStatusSummary(Status status);
+public class StatusEventPublisher {
+    private EventSystem eventSystem;
 
-    void publishIssue(DetectIssue issue);
+    public StatusEventPublisher(EventSystem eventSystem) {
+        this.eventSystem = eventSystem;
+    }
 
-    void publishDetectResult(DetectResult detectResult);
+    public void publishStatusSummary(Status status) {
+        eventSystem.publishEvent(Event.StatusSummary, status);
+    }
+
+    public void publishIssue(DetectIssue issue) {
+        eventSystem.publishEvent(Event.Issue, issue);
+    }
+
+    public void publishDetectResult(DetectResult detectResult) {
+        eventSystem.publishEvent(Event.ResultProduced, detectResult);
+    }
 }
